moved validation from controller to separate class, was making problems for swagger.
